+++ backend/service/base_api/region/find_region_all_logic.v
@@ -54,7 +54,7 @@ pub struct RegionListReq {
 	domain_name          string @[json: 'domainName']
 	continent_code       string @[json: 'continentCode']
 	coord_bounds         string @[json: 'coordBounds']
-	status               []u8   @[json: 'status']
+	status               []i16  @[json: 'status']
 	name_en              string @[json: 'nameEn']
 	name_zh              string @[json: 'nameZh']
 }
@@ -78,7 +78,7 @@ pub struct RegionData {
 	continent_code       ?string @[json: 'continentCode']
 	coord_bounds         ?string @[json: 'coordBounds']
 	sort                 ?int    @[json: 'sort']
-	status               u8      @[json: 'status']
+	status               i16     @[json: 'status']
 	name_en              ?string @[json: 'nameEn']
 	name_zh              ?string @[json: 'nameZh']
 	updater_id           ?string @[json: 'updaterId']
